* src/window.c (_XOPEN_SOURCE): new define
* src/screen.c (is_a_root_window): w is unsigned int
(init_screen): typecast _net_wm_pid as unsigned char*.

* src/main.c (read_rc_file): linesize is size_t

* src/frame.c (frame_read): use a temp var when reading the
:dedicated value.

* src/bar.c (max_line_length): i and start are size_t
(draw_string): i is size_t
(get_mark_box): mark_start and mark_end are size_t

* src/actions.c (find_keybinding): state is unsigned int
(remove_keybinding): likewise
(frame_selector): n is unsigned int
(frame_selector_match): i is size_t. typecast comparison as size_t.

+++ src/actions.c
@@ -455,7 +455,7 @@ find_keybinding_by_action (char *action, rp_keymap *map)
 }
 
 rp_action*
-find_keybinding (KeySym keysym, int state, rp_keymap *map)
+find_keybinding (KeySym keysym, unsigned int state, rp_keymap *map)
 {
   int i;
   for (i = 0; i < map->actions_last; i++)
@@ -536,7 +536,7 @@ replace_keybinding (rp_action *key_action, char *newcmd)
 }
 
 static int
-remove_keybinding (KeySym keysym, int state, rp_keymap *map)
+remove_keybinding (KeySym keysym, unsigned int state, rp_keymap *map)
 {
   int i;
   int found = -1;
@@ -1352,7 +1352,7 @@ cmd_version (int interactive, struct cmdarg **args)
 }
 
 static char *
-frame_selector (int n)
+frame_selector (unsigned int n)
 {
   if (n < strlen (defaults.frame_selectors))
     {
@@ -1368,7 +1368,7 @@ frame_selector (int n)
 static int
 frame_selector_match (char ch)
 {
-  int i;
+  size_t i;
 
   /* Is it in the frame selector string? */
   for (i=0; i<strlen (defaults.frame_selectors); i++)
@@ -1380,7 +1380,7 @@ frame_selector_match (char ch)
   /* Maybe it's a number less than 9 and the frame selector doesn't
      define that many selectors. */
   if (ch >= '0' && ch <= '9'
-      && ch - '0' >= strlen (defaults.frame_selectors))
+      && (size_t)(ch - '0') >= strlen (defaults.frame_selectors))
     {
       return ch - '0';
     }
